Reimplement the car movement.
The movement of cars is now implemented through "Path" classes that represent the trajectory that the car will travel over a road tile given an entry and exit side. There are two types of path: straight and arc (a 90 degree turn). Path classes have an "advance" method that take the distance moved (speed * time) and compute the sprite's position and orientation as it follows the trajectory.